update: file componentize
