+++ src/App.css
@@ -44,218 +44,4 @@ body {
 }
 
 section {
-  width: 100%;
-  min-height: 100vh;
-  display: flex;
-  position: relative;
-}
-
-.intro {
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  overflow: hidden;
-}
-
-.intro-title-wrapper {
-  margin: 0;
-  font-size: 6.25rem;
-  color: #8bc34a;
-  font-family: "CookieRunBlack";
-}
-.intro-title-wrapper p {
-  display: inline-block;
-}
-
-.intro-subTitle {
-  margin: 0;
-  font-size: 1.3rem;
-  color: #f5f5f5;
-  font-family: "CookieRunBold";
-}
-
-.aboutcarrer {
-  margin-top: 10rem;
-  display: flex;
-  flex-direction: column;
-  height: auto;
-  gap: 3rem;
-  align-items: center;
-}
-
-.about {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 1.8rem;
-}
-
-.about-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  width: 30%;
-  gap: 1.8rem;
-}
-
-.about-sentence {
-  margin: 1rem;
-}
-
-.about-sentence h2 {
-  margin-bottom: 2rem;
-  font-size: 2rem;
-  color: #8bc34a;
-  font-family: "CookieRunBold";
-}
-
-.about-sentence p {
-  font-size: 1.5rem;
-  line-height: 2rem;
-}
-
-.profile {
-  width: 20%;
-  height: auto;
-  object-fit: cover;
-  border-radius: 1.2rem;
-}
-
-.career {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 1.5rem;
-}
-
-.career h1 {
-  font-size: 3rem;
-  font-family: "CookieRunBlack";
-}
-
-.career h2 {
-  color: #8bc34a;
-  font-family: "CookieRunBold";
-}
-
-.career p {
-  color: #b3b3b3;
-  line-height: 1.5rem;
-}
-
-.career-badge {
-  display: flex;
-  gap: 0.5rem;
-}
-
-.career-badge h3 {
-  font-family: "CookieRunBold";
-  background-color: rgba(140, 140, 140, 0.5);
-  padding: 0.1rem 0.5rem;
-  border-radius: 0.2rem;
-}
-
-.skills {
-  justify-content: center;
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.skills h1 {
-  font-size: 3rem;
-  font-family: "CookieRunBlack";
-}
-
-.skill-sectionContainer {
-  display: flex;
-  gap: 1rem;
-}
-
-.skill-section {
-  display: flex;
-  flex-direction: column;
-  width: 20rem;
-}
-
-.skill-section h2 {
-  border: 1px solid #8bc34a;
-  background-color: rgba(0, 0, 0, 0.4);
-  color: #8bc34a;
-  font-weight: bold;
-  line-height: 2rem;
-  border-radius: 0.6rem;
-  padding: 2rem;
-  font-family: "CookieRunBold";
-}
-
-.skill-section span {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  background-color: rgba(0, 0, 0, 0.4);
-  gap: 0.5rem;
-  border: 1px solid #f5f5f5;
-  padding: 1rem;
-  margin: 0;
-  border-radius: 0.6rem;
-}
-
-.skill-section span h3 {
-  font-weight: bold;
-  font-family: "CookieRunBold";
-}
-.skill-section span p {
-  color: #ababab;
-}
-.projects {
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.projects h1 {
-  font-size: 3rem;
-  font-family: "CookieRunBlack";
-}
-
-.thanks {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 2rem;
-}
-
-.thanks h1 {
-  font-size: 4rem;
-  font-family: "CookieRunBlack";
-  color: #8bc34a;
-}
-.thanks p {
-  font-size: 2rem;
-}
-
-.thanksLink {
-  display: flex;
-  gap: 1rem;
-}
-
-.thanksLink a {
-  font-size: 2rem;
-  display: block;
-  text-decoration: none;
-  color: #f5f5f5;
-  border: 2px solid #f5f5f5;
-  padding: 1rem 2rem;
-  border-radius: 0.5rem;
-  font-family: "CookieRunBold";
-}
-
-.thanksLink a:hover {
-  border: 2px solid #8bc34a;
-  background-color: #8bc34a;
 }
