+++ src/components/Intro/intro.css
@@ -1,0 +1,27 @@
+.intro {
+  width: 100%;
+  min-height: 100vh;
+  display: flex;
+  position: relative;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  overflow: hidden;
+}
+
+.intro-title-wrapper {
+  margin: 0;
+  font-size: 6.25rem;
+  color: #8bc34a;
+  font-family: "CookieRunBlack";
+}
+.intro-title-wrapper p {
+  display: inline-block;
+}
+
+.intro-subTitle {
+  margin: 0;
+  font-size: 1.3rem;
+  color: #f5f5f5;
+  font-family: "CookieRunBold";
+}
