+++ src/components/Skills/skills.css
@@ -1,0 +1,58 @@
+.skills {
+  width: 100%;
+  min-height: 100vh;
+  display: flex;
+  position: relative;
+  justify-content: center;
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+.skills h1 {
+  font-size: 3rem;
+  font-family: "CookieRunBlack";
+}
+
+.skill-sectionContainer {
+  display: flex;
+  gap: 1rem;
+}
+
+.skill-section {
+  display: flex;
+  flex-direction: column;
+  width: 20rem;
+}
+
+.skill-section h2 {
+  border: 1px solid #8bc34a;
+  background-color: rgba(0, 0, 0, 0.4);
+  color: #8bc34a;
+  font-weight: bold;
+  line-height: 2rem;
+  border-radius: 0.6rem;
+  padding: 2rem;
+  font-family: "CookieRunBold";
+}
+
+.skill-section span {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  background-color: rgba(0, 0, 0, 0.4);
+  gap: 0.5rem;
+  border: 1px solid #f5f5f5;
+  padding: 1rem;
+  margin: 0;
+  border-radius: 0.6rem;
+}
+
+.skill-section span h3 {
+  font-weight: bold;
+  font-family: "CookieRunBold";
+}
+.skill-section span p {
+  color: #ababab;
+}
