+++ src/components/Thanks/thanks.css
@@ -1,0 +1,40 @@
+.thanks {
+  width: 100%;
+  min-height: 100vh;
+  display: flex;
+  position: relative;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 2rem;
+}
+
+.thanks h1 {
+  font-size: 4rem;
+  font-family: "CookieRunBlack";
+  color: #8bc34a;
+}
+.thanks p {
+  font-size: 2rem;
+}
+
+.thanksLink {
+  display: flex;
+  gap: 1rem;
+}
+
+.thanksLink a {
+  font-size: 2rem;
+  display: block;
+  text-decoration: none;
+  color: #f5f5f5;
+  border: 2px solid #f5f5f5;
+  padding: 1rem 2rem;
+  border-radius: 0.5rem;
+  font-family: "CookieRunBold";
+}
+
+.thanksLink a:hover {
+  border: 2px solid #8bc34a;
+  background-color: #8bc34a;
+}
